Relocates the DynamicRender form component attribute type & definition into the component files
