--- src/features/dynamicRender/data/attributes/form/numberInput.ts
@@ -1,5 +1,0 @@
-import { AttributeDefinition } from "features/dynamicRender/types/attributes/definition";
-import { commonInputAttributes } from "./common";
-
-/** The attributes used by number inputs */
-export const numberInputAttributes: AttributeDefinition[] = [...commonInputAttributes];
