--- src/features/dynamicRender/data/attributes/form/textInput.ts
@@ -1,5 +1,0 @@
-import { AttributeDefinition } from "features/dynamicRender/types/attributes/definition";
-import { commonInputAttributes } from "./common";
-
-/** The attributes used by text inputs */
-export const textInputAttributes: AttributeDefinition[] = [...commonInputAttributes];
