--- src/features/dynamicRender/types/attributes/form/checkbox.ts
@@ -1,6 +1,0 @@
-import { CommonInputAttributes } from "./common";
-
-/** Attributes used by the Checkbox element */
-export type CheckboxAttributes = CommonInputAttributes & {
-  value: string;
-};
