--- src/features/dynamicRender/types/attributes/form/numberInput.ts
@@ -1,4 +1,0 @@
-import { CommonInputAttributes } from "./common";
-
-/** Inputs used by the NumberInput element */
-export type NumberInputAttributes = CommonInputAttributes;
