--- src/features/dynamicRender/types/attributes/form/radio.ts
@@ -1,6 +1,0 @@
-import { CommonInputAttributes } from "./common";
-
-/** Attributes used by the Radio element */
-export type RadioAttributes = CommonInputAttributes & {
-  value: string;
-};
