--- src/features/dynamicRender/types/attributes/form/textArea.ts
@@ -1,4 +1,0 @@
-import { CommonInputAttributes } from "./common";
-
-/** Attributes used by the TextArea element */
-export type TextAreaAttributes = CommonInputAttributes;
