--- src/features/dynamicRender/types/attributes/form/textInput.ts
@@ -1,4 +1,0 @@
-import { CommonInputAttributes } from "./common";
-
-/** Attributes used by the TextInput element */
-export type TextInputAttributes = CommonInputAttributes;
